refactor import paths for UserLocationPanel component, removed unused files
