--- venus/src/pages/map/components/locations/PjatkMarker.jsx
@@ -1,9 +1,0 @@
-import { Marker, Popup } from "react-leaflet";
-
-export default function PjatkLocationMarker({ position }) {
-  return (
-    <Marker position={position}>
-      <Popup>PJATK here</Popup>
-    </Marker>
-  );
-}
